Refactor avoidance calc and add tests
